Move stats to their own job (#1324)
