+++ src/lib/jobs/refresh-assistants-counts.ts
